requirements.txt
groq>=0.4.0
fastapi>=0.135.0
uvicorn>=0.42.0
python-dotenv>=1.2.0
httpx>=0.28.0
pydantic>=2.7.0
